backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.7.0
cohere==4.43
groq==0.4.1
pydantic>=2.0
python-dotenv==1.0.0